Upload Storage Service; Add avatar to Company

+++ UTEJob/src/main/java/vn/hcmute/controllers/admin/CompanyController.java
@@ -1,10 +1,16 @@
 package vn.hcmute.controllers.admin;
 
+import java.io.IOException;
+import java.net.http.HttpHeaders;
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 
+import org.codehaus.groovy.util.StringUtil;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.core.io.Resource;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
@@ -14,14 +20,18 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 
+import io.micrometer.common.util.StringUtils;
+import jakarta.persistence.Entity;
 import jakarta.validation.Valid;
 import vn.hcmute.entities.company;
 import vn.hcmute.entities.student;
 import vn.hcmute.entities.users;
 import vn.hcmute.models.companyModel;
 import vn.hcmute.services.ICompanyService;
+import vn.hcmute.services.IStorageService;
 import vn.hcmute.services.IUsersService;
 
 @Controller
@@ -30,6 +40,8 @@ public class CompanyController {
 	@Autowired
 	ICompanyService companyService;
 	IUsersService userService;
+	@Autowired
+	private IStorageService storageService;
 
 	@RequestMapping("")
 	public String list(ModelMap model) {
@@ -54,7 +66,12 @@ public class CompanyController {
 		}
 		company enity = new company();
 		BeanUtils.copyProperties(company, enity);
-	
+		if (!company.getImageFile().isEmpty()) {
+			UUID uuid = UUID.randomUUID();
+			String uuString = uuid.toString();
+			enity.setAvatar(storageService.getStorageFileName(company.getImageFile(), uuString));
+			storageService.store(company.getImageFile(), enity.getAvatar());
+		}
 		companyService.save(enity);
 		String message = "";
 		if (company.getIsEdit() == true) {
@@ -81,11 +98,27 @@ public class CompanyController {
 		model.addAttribute("message", "Công ty không tồn tại");
 		return new ModelAndView("forward:/admin/company", model);
 	}
-
+	
 	@GetMapping("delete/{company_id}")
-	public ModelAndView delete(ModelMap model, @PathVariable("company_id") int company_id) {
-		companyService.deleteById(company_id);
-		model.addAttribute("message", "Xóa thành công");
+	public ModelAndView delete(ModelMap model, @PathVariable("company_id") int company_id) throws IOException {
+		Optional<company> opt = companyService.findById(company_id);
+		if (opt.isPresent()) {
+			if(!org.springframework.util.StringUtils.hasText(opt.get().getAvatar())) {
+				storageService.delete(opt.get().getAvatar());
+			}
+			companyService.deleteById(company_id);
+			model.addAttribute("message", "Xóa thành công");
+		} else {
+			model.addAttribute("message", "Xóa thất bại");
+		}
 		return new ModelAndView("forward:/admin/company", model);
+	}
+	
+	@GetMapping("/images/{filename:.+}")
+	@ResponseBody
+	public ResponseEntity<Resource> serverFile(@PathVariable String filename) {
+		Resource file = storageService.loadAsResource(filename);
+		return ResponseEntity.ok().header(org.springframework.http.HttpHeaders.CONTENT_DISPOSITION,
+				"attachment;filename=\""+ file.getFilename() + "\"").body(file);
 	}
 }
